Split into Corvus.Json.PropertyBag and Corvus.Json.Serialization
