--- Solutions/Corvus.Extensions.System.Text.Json.Specs/Corvus.Extensions.System.Text.Json.Specs.csproj
@@ -1,59 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project Sdk="Microsoft.NET.Sdk">
-  <Import Project="$(EndjinProjectPropsPath)" Condition="$(EndjinProjectPropsPath) != ''" />
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <LangVersion>latest</LangVersion>
-    <Nullable>enable</Nullable>
-
-    <IsPackable>false</IsPackable>
-    <RootNamespace>Corvus.Extensions.Json.Specs</RootNamespace>
-    <!-- Disabling SA1204 because it prioritizes static/non-static over public/non-public, which doesn't fit very well
-         with bindings in SpecFlow.
-         Disabling SA1600, SA1602 (all public types and members to be documented) because test projects need to make lots of types
-         public for technical reasaons, but they are not meant for public consumption, so XML doc comments are only
-         appropriate if they aid understanding within the project.
-    -->
-    <NoWarn>RCS1029;SA0001;SA1204;SA1600;SA1602;CS1591</NoWarn>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <!--
-    This project tests runnable artifacts. To ensure repeatability, we need floating versions in
-    PackageReferences (either in this project, or anywhere in its dependency tree) get resolved to
-    the specific latest matching version. We use floating versions in components to avoid needing
-    to update all of Corvus and Marain any time Microsoft releases a new patch of .NET, but we
-    still need to be deliberate about taking those updates where they emerge at the top of the
-    food chain. This is one of those places, so we use a package lock file to ensure repeatability.
-    The package versions we reference will only change when we explicitly take the change.
-    -->
-    <RestorePackagesWithLockFile>true</RestorePackagesWithLockFile>
-
-    <!--
-    This ensures that during CI builds, the package restore will use exactly what's in the lock
-    file, giving us repeatability. New versions of any of the packages for which we take floating
-    version references (either directly, or implicitly through transitive dependencies) will not
-    be picked up during those builds.
-    When using the project locally, this won't be set, meaning that if any floating references
-    to newer versions become available, they will automatically be acquired, and there will be a
-    record of this happening in source control: the packages.lock.json file will be updated.
-    -->
-    <RestoreLockedMode Condition="'$(ContinuousIntegrationBuild)' == 'true'">true</RestoreLockedMode>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Endjin.RecommendedPractices.GitHub" Version="2.1.4">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Corvus.Testing.SpecFlow.NUnit" Version="2.0.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Corvus.Extensions.System.Text.Json\Corvus.Extensions.System.Text.Json.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Update="specflow.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-</Project>
