--- Solutions/Corvus.Extensions.System.Text.Json.Specs/Samples/ExampleEnum.cs
@@ -1,13 +1,0 @@
-﻿// <copyright file="ExampleEnum.cs" company="Endjin Limited">
-// Copyright (c) Endjin Limited. All rights reserved.
-// </copyright>
-
-namespace Corvus.Extensions.Json.Specs.Samples
-{
-    public enum ExampleEnum
-    {
-        First,
-        Second,
-        Third,
-    }
-}
