+++ Solutions/Corvus.Extensions.System.Text.Json.sln
@@ -10,9 +10,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DevOps", "DevOps", "{C011D1
 		..\GitVersion.yml = ..\GitVersion.yml
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.Extensions.System.Text.Json", "Corvus.Extensions.System.Text.Json\Corvus.Extensions.System.Text.Json.csproj", "{5765619F-5602-41A7-85BA-8078F48076B8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.Json.PropertyBag", "Corvus.Json.PropertyBag\Corvus.Json.PropertyBag.csproj", "{5765619F-5602-41A7-85BA-8078F48076B8}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.Extensions.System.Text.Json.Specs", "Corvus.Extensions.System.Text.Json.Specs\Corvus.Extensions.System.Text.Json.Specs.csproj", "{B0E516F4-3513-4B7C-ADE8-2BBC5D5C5BD5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.Json.Specs", "Corvus.Json.Specs\Corvus.Json.Specs.csproj", "{B0E516F4-3513-4B7C-ADE8-2BBC5D5C5BD5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Corvus.Json.Serialization", "Corvus.Json.Serialization\Corvus.Json.Serialization.csproj", "{0B55B1A4-7BEC-49D6-9A12-CC2DB5B33288}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -28,6 +30,10 @@ Global
 		{B0E516F4-3513-4B7C-ADE8-2BBC5D5C5BD5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B0E516F4-3513-4B7C-ADE8-2BBC5D5C5BD5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B0E516F4-3513-4B7C-ADE8-2BBC5D5C5BD5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0B55B1A4-7BEC-49D6-9A12-CC2DB5B33288}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0B55B1A4-7BEC-49D6-9A12-CC2DB5B33288}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0B55B1A4-7BEC-49D6-9A12-CC2DB5B33288}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0B55B1A4-7BEC-49D6-9A12-CC2DB5B33288}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
