+++ Solutions/Corvus.Json.Specs/Samples/ExampleEnum.cs
@@ -1,0 +1,12 @@
+﻿// <copyright file="ExampleEnum.cs" company="Endjin Limited">
+// Copyright (c) Endjin Limited. All rights reserved.
+// </copyright>
+
+namespace Corvus.Json.Specs.Samples;
+
+public enum ExampleEnum
+{
+    First,
+    Second,
+    Third,
+}
